+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+
+- Add notations for format strings (`Format.t`)
+- `sprintf` and `sscanf` now take a format string of type `Format.t` instead of `string`
+- No longer compatible with 8.8 and 8.9
+
 # 1.0.1
 
 - Add `sscanf`